MyPro/requirements.txt
asgiref==3.3.1
Django==3.1.6
django-postgres-copy==2.4.2
Pillow==8.1.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
xlrd==2.0.1
